refactor: remove for loop
there was only one element in it anyway and at any point we would pop only one book - the book based on which we recommend.

in addition the variable was called "trilogy" - we would actually want to recommend continuation of books (not to mention it would be difficult to group books based on whether they do or do not belong in a saga). E.g. I want to be recommended the two towers if I liked fellowship.